refactor(auth): consolidate JWT guards and remove redundant JwtRefreshGuard
