+++ apps/api/src/modules/auth/auth.controller.ts
@@ -9,7 +9,7 @@ import { CurrentUser } from './decorators/current-user.decorator';
 import { Public } from './decorators/public.decorator';
 import { LoginDto } from './dto/login.dto';
 import { RegisterDto } from './dto/register.dto';
-import { JwtRefreshGuard } from './guards/jwt-refresh.guard';
+import { JwtRefreshGuard } from './guards/jwt-auth.guard';
 
 @Controller('auth')
 export class AuthController {
