+++ apps/api/src/modules/auth/guards/jwt-auth.guard.ts
@@ -20,3 +20,6 @@ export class JwtAuthGuard extends AuthGuard('jwt') {
     return super.canActivate(context);
   }
 }
+
+@Injectable()
+export class JwtRefreshGuard extends AuthGuard('jwt-refresh') {}
