--- apps/api/src/modules/auth/guards/jwt-refresh.guard.ts
@@ -1,5 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class JwtRefreshGuard extends AuthGuard('jwt-refresh') {}
